requirements.txt
amqp==2.4.2
beautifulsoup4==4.6.0
celery==4.2.1
certifi==2019.3.9
chardet==3.0.4
cssselect==1.0.3
decorator==4.3.2
Django==2.1
django-crispy-forms==1.7.0
django-rest-framework==0.1.0
djangorestframework==3.9.0
docx==0.2.4
html5lib==1.0.1
idna==2.8
kombu==4.2.0
lxml==4.3.2
Pillow==5.4.1
PyMuPDF==1.14.10
PyPDF2==1.26.0
python-docx==0.8.10
pytz==2018.3
reportlab==3.5.13
requests==2.18.4
simplejson==3.16.0
six==1.11.0
urllib3==1.22
vine==1.2.0
virtualenv==16.4.3
webencodings==0.5.1
xhtml2pdf==0.2.3
